Added unit testing setup and skeleton (#101)
Nosetests will be installed as a testing requirement. Tests can be run by simply running "nosetests" in the root of the project.

I didn't see a section on actively developing on the abp project, so I added a quick note about the test requirements file in the "Installing from source" section.

I don't typically see a "src" directory in Python applications so I was a bit unsure of where to add the tests directory, but I threw it in there figuring that made the most sense.

I added a simple test to show the testing setup is valid, but that test will need to be flushed out.